+++ src/irmin-pack/pack_value_intf.ml
@@ -99,6 +99,18 @@ module type Sigs = sig
       (Key : Irmin.Key.S with type hash = Hash.t)
       (Commit : Irmin.Commit.Generic_key.S
                   with type node_key = Key.t
-                   and type commit_key = Key.t) :
-    S with type t = Commit.t and type hash = Hash.t and type key = Key.t
+                   and type commit_key = Key.t) : sig
+    include S with type t = Commit.t and type hash = Hash.t and type key = Key.t
+
+    module Commit_direct : sig
+      type address = Offset of int63 | Hash of hash [@@deriving irmin]
+
+      type t = {
+        node_offset : address;
+        parent_offsets : address list;
+        info : Commit.Info.t;
+      }
+      [@@deriving irmin]
+    end
+  end
 end