+++ CHANGES.md
@@ -28,6 +28,10 @@
   - Add a `weight` parameter in the LRU implementation to bound
     memory usage (#2050, @samoht)
 
+- **irmin-tezos-utils**
+  - Add package `irmin-tezos-utils` containing a graphical tool for manual pack
+    files analysis. (#1939, @clecat)
+
 ### Changed
 
 - **irmin**